add multilabel may model; encoding likely breaks downstream
