+++ bento/tests/sample_features_test.py
@@ -11,7 +11,7 @@ class TestSampleFeatures(unittest.TestCase):
         self.assertTrue("cyto_distance_to_cell" in adata.layers)
 
     def test_extract_multicore(self):
-        adata = extract(data, "cyto_distance_to_cell", n_cores=2, copy=True)
+        adata = extract(data, "cyto_distance_to_cell", n_jobs=2, copy=True)
         self.assertTrue("cyto_distance_to_cell" in adata.layers)
 
     def test_CytoDistanceToNucleus(self):
