+++ bento/tools/__init__.py
@@ -5,4 +5,4 @@ from ._locfish_features import (distance_features, moment_stats,
                                 morph_enrichment, nuclear_fraction,
                                 ripley_features)
 from ._sample_features import extract
-from ._spots import PATTERN_NAMES, detect_spots, distr_to_var, spots_diff
+from ._loc_patterns import PATTERN_NAMES, predict_patterns, distr_to_var, spots_diff
