+++ bento/tools/_features.py
@@ -246,6 +246,7 @@ def rasterize_cells(
     global rasterio, torch, torchvision
     if rasterio is None:
         import rasterio
+        from rasterio import features
 
     if torch is None:
         import torch
@@ -274,7 +275,7 @@ def rasterize_cells(
         )
 
         # Rasterize cell
-        base_raster = rasterio.features.rasterize(
+        base_raster = features.rasterize(
             [s],
             fill=0,
             default_value=20,
@@ -284,7 +285,7 @@ def rasterize_cells(
 
         # Rasterize nucleus
         if n is not None:
-            rasterio.features.rasterize(
+            features.rasterize(
                 [n], default_value=40, transform=tf_origin, out=base_raster
             )
 
@@ -320,7 +321,7 @@ def rasterize_cells(
             gene_raster = base_raster.copy()
 
             # Set base as 40
-            gene_raster = rasterio.features.rasterize(
+            gene_raster = features.rasterize(
                 shapes=cg_points.geometry,
                 default_value=40,
                 transform=tf_origin,
@@ -328,7 +329,7 @@ def rasterize_cells(
             )
 
             # Plus 20 per point
-            rasterio.features.rasterize(
+            features.rasterize(
                 shapes=cg_points.geometry,
                 default_value=20,
                 transform=tf_origin,
